requirements.txt

fastapi>=0.110.0
httpx>=0.27.0
langchain-core>=0.3.51
langchain-mcp-adapters==0.0.7
langchain-openai==0.3.12
langgraph==0.3.25
mcp==1.6.0
python-dotenv>=1.0.0
uvicorn>=0.30.0